ECR #333753 - Changing an int to a size_t, so that Bolt can compile with APP SDK v2.7.  There was an ambiguous function call within the buffer class.
[git-p4: depot-paths = "//depot/stg/hsa/drivers/hsa/boltProject/": change = 773150]

+++ BoltForOpenCL/tests/TransformTest/TransformTest.cpp
@@ -224,7 +224,7 @@ void oclTransform(int aSize)
 	std::vector<float> Z0(aSize);
 	std::transform(A.begin(), A.end(), B.begin(), Z0.begin(), boltcl::plus<float>()); // golden answer:
 
-	int bufSize = aSize * sizeof(float);
+	size_t bufSize = aSize * sizeof(float);
 	cl::Buffer bufferA(CL_MEM_READ_ONLY|CL_MEM_USE_HOST_PTR, bufSize, A.data());
 	//cl::Buffer bufferA(begin(A), end(A), true);
 	cl::Buffer bufferB(CL_MEM_READ_ONLY|CL_MEM_USE_HOST_PTR, bufSize, B.data());
